Test classes for each requirement produced.
Changed each requirement class to return an ArrayList containing each
row rather than outputting to the console directly.
User is required to provide login details for each test.
ArrayList from each requirement is compared to an SQL generated
ArrayList.

PerformQuery method added to BaseQuery class.
Removed some unused fields.
Main class deleted - redundant now test classes are produced.

+++ COM1028_Coursework/src/main/java/org/com1028/coursework/OrderDetailsTable.java
@@ -15,7 +15,7 @@ public class OrderDetailsTable {
 			while (rs.next()) {
 				int orderNumber = rs.getInt("orderNumber");
 				int quantityOrdered = rs.getInt("quantityOrdered");
-				int priceEach = rs.getInt("priceEach");
+				double priceEach = rs.getDouble("priceEach");
 				OrderDetails od = new OrderDetails(orderNumber, quantityOrdered, priceEach);
 				this.orderDetailsList.add(od);
 			}
